move CHECK functions to common test file
They can now be reused in other tests.
Switch from err_check to CHECK_OK in srtp_driver.
